+++ TestProjects/CssImport/index.ts
@@ -1,0 +1,3 @@
+import "./some.css";
+
+console.log("Yes just like that");